requirements.txt
openai
gtts
pygame
sounddevice
scipy
python-dotenv